Fix warning in tensor.hpp

+++ src/cuv/basics/tensor.hpp
@@ -1576,7 +1576,7 @@ namespace cuv
                         if (finish<0) finish += o.shape(i);
 #ifndef NDEBUG
                         cuvAssert(finish>start);
-                        cuvAssert(finish <= o.shape(i));
+                        cuvAssert(finish <=(int) o.shape(i));
 #endif
                         t.m_ptr += start*o.stride(i);
                         if(idx.ranges_[i].is_degenerate()){
